+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Shopify-style whitespace control for {{-, -}}, {%-, and -%} delimiters
+- Added built-in support for the Liquid liquid tag, including inline # comments and multiline block parsing
 
 
 ## [0.9.1] - 2026-03-29
@@ -85,6 +86,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Various parsing and rendering improvements
+